% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tech_rep.R
\name{add_tech_rep}
\alias{add_tech_rep}
\title{Add technical replicates}
\usage{
add_tech_rep(se, group)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{group}{Character string vector, or factor, that indicates
the grouping of technical replicates.}
}
\value{
A SummarizedExperiment object, where technical
  replicates have been added in the count matrix.
}
\description{
Add counts across columns of a count matrix corresponding
to technical replicates.
}
\examples{
library(SummarizedExperiment)

cnt <- matrix(1, nrow=3, ncol=9,
                dimnames=list(paste0("g", 1:3), paste0("s", 1:9)))
cdata <- data.frame(group=rep(LETTERS[1:3], each=3), row.names=colnames(cnt))
se <- SummarizedExperiment(assays=list(counts=cnt), colData=cdata)
dim(se)
assays(se)$counts

se <- add_tech_rep(se, se$group)
dim(se)
assays(se)$counts

}
\author{
Robert Castelo
}
